# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Zhongtai API
  version: 0.1.0
servers:
  - url: http://localhost:4000/api
paths:
...
  /ai/tasks:
    get:
      summary: List tasks
      responses:
        "200":
          description: OK
    post:
      summary: Create task (async)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                toolId:
                  type: string
                payload:
                  type: object
              required: [toolId, payload]
      responses:
        "200":
          description: OK
  /ai/tasks/{taskId}:
    get:
      summary: Get task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /ai/tasks/{taskId}/cancel:
    post:
      summary: Cancel task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...